Major refactor of directories and roles, updates to syntax, and bug fixes
This was a large rework of the general structure of the wuvt ansible repo.
As part of the process of adding roles for creating ldap and nfs clients which
enable the new central auth mechanism, it was decided to clean up as well
as update all the existing ansible roles, plays, and associated files. This
committ comprises the largest portion of major revisions that dont directly
contribute new roles, plays, or functionality. The exception to this are the
ldap-client and nfs-client plays.