Move automation to a Docker container
Since Liquidsoap hasn't been updated in the Debian/Ubuntu repositories
for quite some time, we've been running an old version in production.
Instead, bundle everything together in a Docker container so that we can
easily run the latest version of Liquidsoap and more easily manage
dependencies.

For now, the helper queue_underwriting.py and playlist-loader scripts
run as cronjobs outside of the container. In the future, these should
probably be moved to some sort of web service to remove this dependency
and allow for additional flexibility.