Restrict SSH from external IPs
On fukushima and stirling, restrict SSH access from external IPs to
users in the ssh-allowdirect group. If a user is not in this group,
the source IP must be that of the internal WUVT network.

Ideally we'd switch this to use variables for the networks and hosts
that require this, but this works for now.